Refactoring to use the SubGroupChoreography class.
