+++ ChoreographyHive/choreography/group_step.py
@@ -3,6 +3,7 @@ from typing import Callable, List
 from rlbot.agents.base_agent import SimpleControllerState
 from rlbot.utils.structures.game_data_struct import GameTickPacket
 
+from choreography.choreography import ChoreographyBase
 from choreography.drone import Drone
 
 
@@ -70,3 +71,52 @@ class BlindBehaviorStep(PerDroneStep):
     def blind(self, packet: GameTickPacket, drone: Drone, elapsed: float):
         drone.ctrl = self.controls
         return StepResult(finished=False)
+
+
+class SubGroupChoreography(ChoreographyBase):
+    """
+    This defines a sub-group of drones that may behave differently from the other drones present
+    on the field.
+    """
+    def __init__(self, drones: List[Drone], start_time: float):
+        super().__init__()
+        self.drones = drones
+        self.start_time = start_time
+
+
+class SubGroupOrchestrator(GroupStep):
+    def __init__(self, group_list: List[SubGroupChoreography]):
+        self.group_list = group_list
+        self.active_groups: List[SubGroupChoreography] = []
+        self.completed = False
+        self.group_list.sort(key=lambda x: x.start_time)
+        self.start_time = None
+
+    def update(self, current_time):
+        if not self.start_time:
+            self.start_time = current_time
+
+        if len(self.group_list) == 0 and len(self.active_groups) == 0:
+            self.completed = True
+            return
+        self.active_groups = [g for g in self.active_groups if not g.finished]
+        num_consumed = 0
+        elapsed_time = current_time - self.start_time
+        for candidate in self.group_list:
+            if candidate.start_time > elapsed_time:
+                break
+            candidate.generate_sequence(candidate.drones)
+            self.active_groups.append(candidate)
+            num_consumed += 1
+
+        self.group_list = self.group_list[num_consumed:]
+
+    def step(self, packet: GameTickPacket):
+        for choreo in self.active_groups:
+            # Pass the choreo its own drones. Don't make it aware of all the drones in the match.
+            choreo.step(packet, choreo.drones)
+
+    def perform(self, packet: GameTickPacket, drones: List[Drone]) -> StepResult:
+        self.update(packet.game_info.seconds_elapsed)
+        self.step(packet)
+        return StepResult(finished=self.completed)
